// src/app/student&staff/model/staff.model.ts
export class Staff {
 address?: string;
  basic_PAY?: number;
  city?: string;
  dept?: string;
  dob?: string;
  doj?: string;
  email?: string;
  empname?: string;
  empno?: number;
  m_STATUS?: string;
  mobile?: string;
  nature_OF_JOB?: string;
  phone?: string;
  pin?: string;
  sex?: string;
  state?: string;
}